requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.10
pytesseract>=0.3.8
Pillow>=9.0.0
transformers>=4.20.0
torch>=1.10.0
pytest>=7.0.0
pytest-cov>=4.0.0 